+++ partiql-planner/src/main/kotlin/org/partiql/planner/internal/Env.kt
@@ -15,19 +15,16 @@ import org.partiql.planner.internal.typer.FnResolver
 import org.partiql.spi.BindingCase
 import org.partiql.spi.BindingName
 import org.partiql.spi.BindingPath
-import org.partiql.spi.Plugin
-import org.partiql.spi.connector.Connector
 import org.partiql.spi.connector.ConnectorMetadata
 import org.partiql.spi.connector.ConnectorObjectHandle
 import org.partiql.spi.connector.ConnectorObjectPath
 import org.partiql.spi.connector.ConnectorSession
-import org.partiql.spi.connector.Constants
 import org.partiql.types.StaticType
 import org.partiql.types.StructType
 import org.partiql.types.TupleConstraint
 
 /**
- * Handle for associating a catalog with the metadata; pair of catalog to data.
+ * Handle for associating a catalog name with catalog related metadata objects.
  */
 internal typealias Handle<T> = Pair<String, T>
 
@@ -124,12 +121,12 @@ internal enum class ResolutionStrategy {
  * PartiQL Planner Global Environment of Catalogs backed by given plugins.
  *
  * @property headers        List of namespaced definitions
- * @property plugins        List of plugins for global resolution
+ * @property catalogs       List of plugins for global resolution
  * @property session        Session details
  */
 internal class Env(
     private val headers: List<Header>,
-    private val plugins: List<Plugin>,
+    private val catalogs: Map<String, ConnectorMetadata>,
     private val session: PartiQLPlanner.Session,
 ) {
 
@@ -146,26 +143,6 @@ internal class Env(
     private val connectorSession = object : ConnectorSession {
         override fun getQueryId(): String = session.queryId
         override fun getUserId(): String = session.userId
-    }
-
-    /**
-     * Map of catalog names to its underlying connector
-     */
-    private val catalogs: Map<String, Connector>
-
-    // Initialize connectors
-    init {
-        val catalogs = mutableMapOf<String, Connector>()
-        val connectors = plugins.flatMap { it.getConnectorFactories() }
-        // map catalogs to connectors
-        for ((catalog, config) in session.catalogConfig) {
-            // find corresponding connector
-            val connectorName = config[Constants.CONFIG_KEY_CONNECTOR_NAME].stringValue
-            val connector = connectors.first { it.getName() == connectorName }
-            // initialize connector with given config
-            catalogs[catalog] = connector.create(catalog, config)
-        }
-        this.catalogs = catalogs.toMap()
     }
 
     /**
@@ -212,8 +189,7 @@ internal class Env(
      */
     private fun getMetadata(catalogName: BindingName): Handle<ConnectorMetadata>? {
         val catalogKey = catalogs.keys.firstOrNull { catalogName.isEquivalentTo(it) } ?: return null
-        val connector = catalogs[catalogKey] ?: return null
-        val metadata = connector.getMetadata(connectorSession)
+        val metadata = catalogs[catalogKey] ?: return null
         return catalogKey to metadata
     }