+++ plugins/partiql-memory/src/main/kotlin/org/partiql/plugins/memory/MemoryPlugin.kt
@@ -5,9 +5,10 @@ import org.partiql.spi.connector.Connector
 import org.partiql.spi.function.PartiQLFunction
 import org.partiql.spi.function.PartiQLFunctionExperimental
 
-class MemoryPlugin(val provider: MemoryCatalog.Provider) : Plugin {
-    override fun getConnectorFactories(): List<Connector.Factory> = listOf(MemoryConnector.Factory(provider))
+class MemoryPlugin(private val catalogs: Map<String, MemoryConnector>) : Plugin {
 
-    @PartiQLFunctionExperimental
-    override fun getFunctions(): List<PartiQLFunction> = emptyList()
+    override val factory: Connector.Factory = MemoryConnector.Factory(catalogs)
+
+    @OptIn(PartiQLFunctionExperimental::class)
+    override val functions: List<PartiQLFunction> = emptyList()
 }